Develop Parsing API Payloads. (#7)
* Add properties and initializer to BackendController.

Added loggedInUser where the current main user will be stored. Created dictionaries for users, properties, and pickups where they will be stored with their IDs ad Key. Parsers dictionary will hold closures to parse API Response dynamically.

* Implement propertyParser closure.

* Implement propertiesParser closure.

* Implement userParser closure.

* Implement pickupParser closure.

* Populate parsers dictionary property with all the closures.

* Add stats array of ImpactStats to Property class.

* Update queryAPI.

Unfortunately this implementation doesn't properly work for impactStats case.

* Dynamic parsing accounts for payload without a parser.

* Fix stats property. Expand initializer to parse impact stats from data.

* Add ImpactStats to property queries.

* Meaningful error messages for User initializer.

* Meaningful error messages for Property initializer.

* Meaningful error messages for Pickup initializer.

* Remove statsById method.

This was the original function used for testing. No longer needed.

* Implement convenience functions.

Each one of these functions will call fetchAPI with the corresponding enum and the passed in IDs. For impactStats, it will handle checking for the property whose id was used for fetching and instantiating and storing an ImpactStats object to the corresponding property.